+++ tests/integration/cases/dict_mixed_scalars/D_heterogeneous_strategy_record_combined@v2.d
@@ -1,0 +1,11 @@
+struct Record0 { long a; string b; }
+void main() {
+auto my_data = Record0(
+    1,
+    "x",
+);
+my_data = Record0(
+    1,
+    "x",
+);
+}